+++ extension-androidauto/src/main/java/com/mapbox/maps/extension/androidauto/MapboxCarMapObserver.java
@@ -6,13 +6,11 @@ import androidx.annotation.NonNull;
 import androidx.car.app.SurfaceCallback;
 
 import com.mapbox.maps.EdgeInsets;
-import com.mapbox.maps.MapboxExperimental;
 
 /**
  * Many downstream services will not work until the surface has been created and the map has
  * loaded. This interface allows you to create custom Mapbox experiences for the car.
  */
-@MapboxExperimental
 public interface MapboxCarMapObserver {
 
   /**