+++ extension-androidauto/src/main/java/com/mapbox/maps/extension/androidauto/MapboxCarMapSurface.kt
@@ -3,7 +3,6 @@ package com.mapbox.maps.extension.androidauto
 import androidx.car.app.CarContext
 import androidx.car.app.SurfaceContainer
 import com.mapbox.maps.MapSurface
-import com.mapbox.maps.MapboxExperimental
 
 /**
  * This contains the Android Auto head unit map information.
@@ -13,7 +12,6 @@ import com.mapbox.maps.MapboxExperimental
  * @property mapSurface Mapbox controllable interface
  * @property surfaceContainer A container for the Surface created by the car.
  */
-@MapboxExperimental
 class MapboxCarMapSurface internal constructor(
   val carContext: CarContext,
   val mapSurface: MapSurface,